Interfaces
A basic code explaining the Abstract class in Java.
Here Vehicle is an abstract class which is like an idea.
Class Can and Scooter extends the abstract class Vehicle.
The abstract methods declared in the class Vehicle must be defined or implemented in the class those inherit the Abstract class. Thus, Car and Scooter classes have implemented the methods declared as abstract I the class Vehicle.