Create InnerClasses
The outer class here i.e., InnerClass has the main() method. This class have two inner classes Toy which is a public class and Clothes which is a private class.

Let's think the inner classes as variable or methods of the outer class. If the inner classes are static then we don't nee to create objects of it as they will be the members of the outer class.

In this code, our inner classes are not static. Hence, we have to create the object of outer class first and then using the outer class object we have created the object toy of inner class Toy.